student-report-synthesis/requirements.txt
# Project dependencies
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
openai>=1.0.0
reportlab==3.6.12
pillow==9.5.0
python-docx==0.8.11
python-dotenv==1.0.0
jinja2==3.1.2
xhtml2pdf==0.2.11
